requirements-frozen.txt
aiodns==3.0.0
aiosmtpd==1.4.4.post2
alembic==1.1.0
aniso8601==7.0.0
atpublic==3.1.1
attrs==19.1.0
azure-common==1.1.28
azure-core==1.26.3
azure-servicebus==7.8.2
barnum==0.5.1
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cssselect==1.2.0
cssutils==2.6.0
emails==0.5.15
et-xmlfile==1.1.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
flask-emails==0.4.2
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
future==0.17.1
idna==2.8
importlib-metadata==6.0.0
importlib-resources==5.10.2
iso3166==2.1.1
isodate==0.6.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jsonschema==3.0.2
luqum==0.8.0
lxml==4.5.0
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.0.1
marshmallow-sqlalchemy==0.17.2
minio==4.0.21
msrest==0.7.1
oauthlib==3.2.2
openpyxl==2.6.4
pdfkit==1.0.0
Pillow==9.4.0
pillow-heif==0.1.4
ply==3.11
premailer==3.6.1
psycopg2==2.8.3
pycares==4.3.0
pycountry==22.3.5
pycparser==2.19
PyJWT==1.7.1
PyPDF2==2.10.5
pyrsistent==0.15.4
python-dateutil==2.8.0
python-editor==1.0.4
python-magic==0.4.27
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.1
schwifty==2022.9.0
six==1.12.0
SQLAlchemy==1.3.8
SQLAlchemy-Continuum==1.3.9
SQLAlchemy-Utils==0.34.2
typing_extensions==4.4.0
uamqp==1.6.3
urllib3==1.25.3
verify-email==2.4.3
Werkzeug==0.15.5
zipp==3.12.1
